+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Forge Recommended Versioning](https://mcforge.readthedocs.io/en/latest/conventions/versioning/).
 
+## [1.20.4-3.1.3.0] - 2024-02-20
+### Added
+- new recipe type "bucketlib:bucket_filling_shaped" to specify crafting recipes that fill buckets
+- new recipe type "bucketlib:bucket_filling_shapeless" to specify crafting recipes that fill buckets
+
+### Fixed
+- ingredients of type "bucketlib:empty" detected some not intended items
+
 ## [1.20.4-3.1.2.1] - 2024-02-12
 ### Fixed
 - startup crash when starting in Forge environment was fixed (thanks to Ellesmera for the report)